Payment entities renamed.
